refactor(plugin-ember-octane) move types into their own directory
